refactor(client): alters NodeFormAttributes to be more generic so that it can be used by other entities with attributes
