+++ src/client/src/assets/locales/en/translation.json
@@ -89,7 +89,7 @@
     "remove": "Remove attribute",
     "dialog": {
       "title": "Select attributes(s)",
-      "description": "The attributes listed below are predefined and can be reused across multiple nodes.",
+      "description": "The attributes that are listed below can be reused across multiple entities.",
       "search": "Search",
       "add": "Add"
     }
