--- src/client/src/content/forms/node/attributes/NodeFormAttributes.helpers.tsx
@@ -1,29 +1,0 @@
-import { Aspect, AttributeLibCm } from "@mimirorg/typelibrary-types";
-import { UseFieldArrayReturn } from "react-hook-form";
-import { mapAttributeLibCmsToInfoItems } from "../../../../utils/mappers";
-import { FormNodeLib } from "../types/formNodeLib";
-
-export const onAddAttributes = (
-  ids: string[],
-  attributeFields: UseFieldArrayReturn<FormNodeLib, "attributeIdList">
-) => {
-  ids.forEach((id) => {
-    const attributeHasNotBeenAdded = !attributeFields.fields.some((f) => f.value === id);
-    if (attributeHasNotBeenAdded) {
-      attributeFields.append({ value: id });
-    }
-  });
-};
-
-export const getSelectItemsFromAttributeLibCms = (attributes?: AttributeLibCm[]) => {
-  if (!attributes || attributes.length == 0) return [];
-
-  return mapAttributeLibCmsToInfoItems(attributes);
-};
-
-export const prepareAttributes = (attributes?: AttributeLibCm[], aspects?: Aspect[]) => {
-  if (!attributes || attributes.length == 0) return [];
-  if (!aspects || aspects.length == 0) return [];
-
-  return attributes.filter((a) => aspects.some((x) => x === a.aspect)).sort((a, b) => a.discipline - b.discipline);
-};
