+++ src/client/src/content/forms/node/variants/LocationNode.tsx
@@ -1,6 +1,5 @@
 import { Aspect } from "@mimirorg/typelibrary-types";
 import { Control, UseFormRegister } from "react-hook-form";
-import { NodeFormAttributes } from "../attributes/NodeFormAttributes";
 import { NodeFormPredefinedAttributes } from "../predefined-attributes/NodeFormPredefinedAttributes";
 import { FormNodeLib } from "../types/formNodeLib";
 
@@ -11,7 +10,6 @@ export interface LocationNodeProps {
 
 export const LocationNode = ({ control, register }: LocationNodeProps) => (
   <>
-    <NodeFormAttributes control={control} register={register} aspects={[Aspect.Location]} />
     <NodeFormPredefinedAttributes control={control} register={register} aspects={[Aspect.Location]} />
   </>
 );
